+++ pkg/opensearch/opensearch.go
@@ -1,15 +1,20 @@
 package opensearch
 
 import (
+	"bytes"
 	"context"
-	"errors"
 	"fmt"
+	"io"
 	"net/http"
+	"net/url"
+	"path"
+	"strings"
 
 	"github.com/bitly/go-simplejson"
 	"github.com/grafana/grafana-plugin-sdk-go/backend"
 	"github.com/grafana/grafana-plugin-sdk-go/backend/instancemgmt"
 	"github.com/grafana/grafana-plugin-sdk-go/backend/log"
+	"github.com/grafana/grafana-plugin-sdk-go/experimental/errorsource"
 	"github.com/grafana/opensearch-datasource/pkg/opensearch/client"
 )
 
@@ -60,25 +65,25 @@ func (ds *OpenSearchDatasource) QueryData(ctx context.Context, req *backend.Quer
 		return nil, err
 	}
 
-	errRefID, err := handleServiceMapPrefetch(ctx, osClient, req)
-	if err != nil {
-		return wrapServiceMapPrefetchError(errRefID, err)
+	response := handleServiceMapPrefetch(ctx, osClient, req)
+	if response != nil {
+		return response, nil
 	}
 
 	query := newQueryRequest(osClient, req.Queries, req.PluginContext.DataSourceInstanceSettings)
-	response, err := wrapError(query.execute(ctx))
+	response, err = wrapError(query.execute(ctx))
 	return response, err
 }
 
 // handleServiceMapPrefetch inspects the given request, and, if it wants a serviceMap, creates and
 // calls the Prefetch query to get the services and operations lists that are required for
 // the associated Stats query. It then adds these parameters to the originating query so
-// the Stats query can be created later.
-func handleServiceMapPrefetch(ctx context.Context, osClient client.Client, req *backend.QueryDataRequest) (string, error) {
+// the Stats query can be created later. Returns a response with an error if the request fails.
+func handleServiceMapPrefetch(ctx context.Context, osClient client.Client, req *backend.QueryDataRequest) *backend.QueryDataResponse {
 	for i, query := range req.Queries {
 		model, err := simplejson.NewJson(query.JSON)
 		if err != nil {
-			return "", err
+			return wrapServiceMapPrefetchError(query.RefID, err)
 		}
 		queryType := model.Get("queryType").MustString()
 		luceneQueryType := model.Get("luceneQueryType").MustString()
@@ -88,7 +93,9 @@ func handleServiceMapPrefetch(ctx context.Context, osClient client.Client, req *
 			q := newQueryRequest(osClient, []backend.DataQuery{prefetchQuery}, req.PluginContext.DataSourceInstanceSettings)
 			response, err := q.execute(ctx)
 			if err != nil {
-				return query.RefID, err
+				return wrapServiceMapPrefetchError(query.RefID, err)
+			} else if response.Responses[query.RefID].Error != nil {
+				return wrapServiceMapPrefetchError(query.RefID, response.Responses[query.RefID].Error)
 			}
 			services, operations := extractParametersFromServiceMapFrames(response)
 
@@ -99,41 +106,29 @@ func handleServiceMapPrefetch(ctx context.Context, osClient client.Client, req *
 			// An error here _should_ be impossible but since services and operations are coming from outside,
 			// handle it just in case
 			if err != nil {
-				return query.RefID, err
+				return wrapServiceMapPrefetchError(query.RefID, err)
 			}
 			req.Queries[i].JSON = newJson
-			return "", nil
+			return nil
 		}
 	}
-	return "", nil
+	return nil
 }
 
-func wrapServiceMapPrefetchError(refId string, err error) (*backend.QueryDataResponse, error) {
-	if refId != "" {
-		return &backend.QueryDataResponse{
-			Responses: map[string]backend.DataResponse{
-				refId: {
-					Error: fmt.Errorf(`Error fetching service map info: %w`, err),
-				}},
-		}, nil
+func wrapServiceMapPrefetchError(refId string, err error) *backend.QueryDataResponse {
+	if err != nil {
+		response := backend.NewQueryDataResponse()
+		err = errorsource.PluginError(err, false) // keeps downstream source if present
+		err = fmt.Errorf(`Error fetching service map info: %w`, err)
+		return errorsource.AddErrorToResponse(refId, response, err)
 	}
-	return nil, err
+	return nil
 }
 
 func wrapError(response *backend.QueryDataResponse, err error) (*backend.QueryDataResponse, error) {
-	var invalidQueryTypeError invalidQueryTypeError
-	if errors.As(err, &invalidQueryTypeError) {
-		return &backend.QueryDataResponse{
-			Responses: map[string]backend.DataResponse{
-				invalidQueryTypeError.refId: {
-					Error: fmt.Errorf(`%w, expected Lucene or PPL`, err),
-				}},
-		}, nil
-	}
 	if err != nil {
 		return response, fmt.Errorf("OpenSearch data source error: %w", err)
 	}
-
 	return response, err
 }
 
@@ -175,4 +170,65 @@ func extractParametersFromServiceMapFrames(resp *backend.QueryDataResponse) ([]s
 		}
 	}
 	return services, operations
+}
+
+func (ds *OpenSearchDatasource) CallResource(ctx context.Context, req *backend.CallResourceRequest, sender backend.CallResourceResponseSender) error {
+	// allowed paths for resource calls:
+	// - empty string for fetching db version
+	// - /_mapping for fetching index mapping, e.g. requests going to `index/_mapping`
+	// - _msearch for executing getTerms queries
+	// - _mapping for fetching "root" index mappings
+	if req.Path != "" && !strings.HasSuffix(req.Path, "/_mapping") && req.Path != "_msearch" && req.Path != "_mapping" {
+		return fmt.Errorf("invalid resource URL: %s", req.Path)
+	}
+
+	osUrl, err := createOpensearchURL(req, req.PluginContext.DataSourceInstanceSettings.URL)
+	if err != nil {
+		return err
+	}
+
+	request, err := http.NewRequestWithContext(ctx, req.Method, osUrl, bytes.NewBuffer(req.Body))
+	if err != nil {
+		return err
+	}
+
+	response, err := ds.HttpClient.Do(request)
+	if err != nil {
+		return err
+	}
+
+	body, err := io.ReadAll(response.Body)
+	if err != nil {
+		return err
+	}
+
+	responseHeaders := map[string][]string{
+		"content-type": {"application/json"},
+	}
+
+	if encoding := response.Header.Get("Content-Encoding"); encoding != "" {
+		responseHeaders["content-encoding"] = []string{response.Header.Get("Content-Encoding")}
+	}
+
+	return sender.Send(&backend.CallResourceResponse{
+		Status:  response.StatusCode,
+		Headers: responseHeaders,
+		Body:    body,
+	})
+}
+
+func createOpensearchURL(req *backend.CallResourceRequest, urlStr string) (string, error) {
+	osUrl, err := url.Parse(urlStr)
+	if err != nil {
+		return "", fmt.Errorf("failed to parse data source URL: %s, error: %w", urlStr, err)
+	}
+	osUrl.Path = path.Join(osUrl.Path, req.Path)
+	osUrlString := osUrl.String()
+	// If the request path is empty and the URL does not end with a slash, add a slash to the URL.
+	// This ensures that for version checks executed to the root URL, the URL ends with a slash.
+	// This is helpful, for example, for load balancers that expect URLs to match the pattern /.*.
+	if req.Path == "" && osUrlString[len(osUrlString)-1:] != "/" {
+		return osUrl.String() + "/", nil
+	}
+	return osUrlString, nil
 }