+++ pkg/opensearch/ppl_response_parser.go
@@ -8,6 +8,7 @@ import (
 	simplejson "github.com/bitly/go-simplejson"
 	"github.com/grafana/grafana-plugin-sdk-go/backend"
 	"github.com/grafana/grafana-plugin-sdk-go/data"
+	"github.com/grafana/grafana-plugin-sdk-go/experimental/errorsource"
 	"github.com/grafana/opensearch-datasource/pkg/null"
 	"github.com/grafana/opensearch-datasource/pkg/opensearch/client"
 	"github.com/grafana/opensearch-datasource/pkg/utils"
@@ -37,16 +38,14 @@ func (rp *pplResponseParser) parseResponse(configuredFields client.ConfiguredFie
 	}
 
 	if rp.Response.Error != nil {
-		return &backend.DataResponse{
-			Error: getErrorFromPPLResponse(rp.Response),
-			Frames: []*data.Frame{
-				{
-					Meta: &data.FrameMeta{
-						Custom: debugInfo,
-					},
+		errResp := errorsource.Response(errorsource.DownstreamError(getErrorFromPPLResponse(rp.Response), false))
+		errResp.Frames = []*data.Frame{
+			{
+				Meta: &data.FrameMeta{
+					Custom: debugInfo,
 				},
-			},
-		}, nil
+			}}
+		return &errResp, nil
 	}
 
 	queryRes := &backend.DataResponse{
@@ -91,7 +90,8 @@ func (rp *pplResponseParser) parsePPLResponse(queryRes *backend.DataResponse, co
 				}
 				ts, err := rp.parseTimestamp(row[fieldIdx], timestampFormat)
 				if err != nil {
-					return nil, err
+					errResp := errorsource.Response(errorsource.PluginError(err, false))
+					return &errResp, nil
 				}
 				value = *utils.NullFloatToNullableTime(ts)
 			}
@@ -139,7 +139,8 @@ func (rp *pplResponseParser) parsePPLResponse(queryRes *backend.DataResponse, co
 func (rp *pplResponseParser) parseTimeSeries(queryRes *backend.DataResponse) (*backend.DataResponse, error) {
 	t, err := getTimeSeriesResponseMeta(rp.Response.Schema)
 	if err != nil {
-		return nil, err
+		errResp := errorsource.Response(errorsource.PluginError(err, false))
+		return &errResp, nil
 	}
 
 	valueName := rp.getSeriesName(t.valueIndex)
@@ -152,7 +153,8 @@ func (rp *pplResponseParser) parseTimeSeries(queryRes *backend.DataResponse) (*b
 	for i, datarow := range rp.Response.Datarows {
 		err := rp.addDatarow(newFrame, i, datarow, t)
 		if err != nil {
-			return nil, err
+			errResp := errorsource.Response(errorsource.PluginError(err, false))
+			return &errResp, nil
 		}
 	}